Fixes #1257
The editor didn't test that a mission area actually existed, so when you switched to the editor, it would trigger the crash.

It now tests if there is a mission area when switching to the editor tool, and if there is not, prompts for the creation of a new one, similar to the auto-prompt for creating terrain blocks.